reorganize request details files
