+++ tests/__tests__/Components/ReqDetail/headers/RequestPayload.spec.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 
-import RequestPayload from '../../../../../src/Components/ReqDetail/sections/RequestPayload';
+import RequestPayload from '../../../../../src/Components/ReqDetail/request/RequestPayload';
 
 describe('RequestPayload', () => {
   const props = {
